# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "企业项目知识库平台",
    "description": "Phase 1 + Phase 2 完整 API:\n\n**Phase 1:**\n- 多模态文档转换管道 (PDF/Word/Excel/PPT/图片/音视频/网页 → Markdown)\n- 基础 RAG 检索 (BGE-M3 三合一 + Qdrant + RRF + Reranker)\n- MCP 工具化查询 (list/outline/search/read 四工具)\n- 引用溯源 (内联引用标记 + Markdown 章节跳转 + 源文件定位)\n\n**Phase 2:**\n- Agentic RAG (LangGraph Plan-and-Execute, 最多4步, 推理透明化)\n- Skills 体系 (文档分析/跨文档对比/项目健康/报告生成)\n- SSE 流式输出 (GET /api/v1/search/stream, 感知延迟↓80%)\n- Contextual Retrieval (chunk 上下文前缀, 检索失败↓67%)",
    "version": "3.0.0"
  },
  "paths": {
...
    "/api/v1/search/": {
      "post": {
        "tags": [
          "Search"
        ],
        "summary": "Search",
        "description": "RAG 检索主入口（非流式）\n\n完整 11 步管道:\n1. 查询输入  2. 术语识别  3. 查询路由\n4-7. 混合检索 + RRF + Reranker\n8. Agentic 路径判断  9. 术语注入\n10. 答案生成 + 引用  11. 审计日志",
        "operationId": "search_api_v1_search__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/search/stream": {
      "get": {
        "tags": [
          "Search"
        ],
        "summary": "Search Stream",
        "description": "SSE 流式检索 — 实时推送检索过程和答案\n\n事件类型:\n  - step:     进度状态更新\n  - thinking: Agentic RAG 推理思考过程\n  - chunk:    LLM 生成的文本增量\n  - citation: 引用来源信息\n  - done:     完成信号 + 统计信息\n  - error:    错误信息",
        "operationId": "search_stream_api_v1_search_stream_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 2000,
              "title": "Query"
            }
          },
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "top_k",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "minimum": 1,
              "default": 5,
              "title": "Top K"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/search/skill/{skill_name}": {
      "post": {
        "tags": [
          "Search"
        ],
        "summary": "Invoke Skill",
        "description": "直接调用指定 Skill\n\n可用 Skills:\n- document_analysis: 单文档分析\n- cross_document_compare: 跨文档对比\n- project_health: 项目健康评估\n- report_generation: 报告生成",
        "operationId": "invoke_skill_api_v1_search_skill__skill_name__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skill_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Skill Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_invoke_skill_api_v1_search_skill__skill_name__post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/search/skills": {
      "get": {
        "tags": [
          "Search"
        ],
        "summary": "List Skills",
        "description": "列出所有已注册的 Skills",
        "operationId": "list_skills_api_v1_search_skills_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_invoke_skill_api_v1_search_skill__skill_name__post": {
        "properties": {
          "request": {
            "$ref": "#/components/schemas/SearchRequest"
          },
          "doc_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "format": "uuid"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doc Ids"
          },
          "params": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Params"
          }
        },
        "type": "object",
        "required": [
          "request"
        ],
        "title": "Body_invoke_skill_api_v1_search_skill__skill_name__post"
      },
...
      "CitationSource": {
        "properties": {
          "doc_id": {
            "type": "string",
            "format": "uuid",
            "title": "Doc Id"
          },
          "doc_name": {
            "type": "string",
            "title": "Doc Name"
          },
          "doc_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doc Type"
          },
          "md_path": {
            "type": "string",
            "title": "Md Path"
          },
          "section_path": {
            "type": "string",
            "title": "Section Path"
          },
          "section_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section Title"
          },
          "source_path": {
            "type": "string",
            "title": "Source Path"
          },
          "source_format": {
            "type": "string",
            "title": "Source Format"
          },
          "page_num": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Num"
          },
          "sheet_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sheet Name"
          },
          "timestamp_sec": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp Sec"
          },
          "content_snippet": {
            "type": "string",
            "title": "Content Snippet"
          },
          "highlight_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Highlight Text"
          },
          "checksum": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checksum"
          },
          "upload_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upload By"
          },
          "upload_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upload At"
          },
          "relevance_score": {
            "type": "number",
            "title": "Relevance Score"
          },
          "retrieval_method": {
            "$ref": "#/components/schemas/RetrievalMethod"
          }
        },
        "type": "object",
        "required": [
          "doc_id",
          "doc_name",
          "md_path",
          "section_path",
          "source_path",
          "source_format",
          "content_snippet",
          "relevance_score",
          "retrieval_method"
        ],
        "title": "CitationSource",
        "description": "引用溯源元数据"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RetrievalMethod": {
        "type": "string",
        "enum": [
          "vector",
          "fulltext",
          "mcp_tool"
        ],
        "title": "RetrievalMethod"
      },
      "SearchRequest": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Query"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "top_k": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Top K",
            "default": 5
          }
        },
        "type": "object",
        "required": [
          "query",
          "project_id"
        ],
        "title": "SearchRequest"
      },
      "SearchResponse": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/CitationSource"
            },
            "type": "array",
            "title": "Citations"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SearchResult"
            },
            "type": "array",
            "title": "Results"
          },
          "retrieval_method": {
            "type": "string",
            "title": "Retrieval Method"
          },
          "latency_ms": {
            "type": "number",
            "title": "Latency Ms"
          }
        },
        "type": "object",
        "required": [
          "query",
          "answer",
          "citations",
          "results",
          "retrieval_method",
          "latency_ms"
        ],
        "title": "SearchResponse",
        "description": "检索响应"
      },
      "SearchResult": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content"
          },
          "citation": {
            "$ref": "#/components/schemas/CitationSource"
          },
          "score": {
            "type": "number",
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "content",
          "citation",
          "score"
        ],
        "title": "SearchResult",
        "description": "单条检索结果"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}